Organised directory (kind of)
It looks a bit better than before.... but it's still messy, but I'll fix it when I actually know how we're doing it...HAHAHAHHA
