+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "client-install": "cd client && npm install",
     "start": "node functions/index.js",
-    "server": "nodemon functions/index.js",
+    "server": "nodemon api/functions/index.js",
     "client": "npm start --prefix client",
     "dev": "concurrently \"npm run server\" \"npm run client\""
   },
